chore: extract schemas + delete notifications_facebook_demo.py (closes #665 cat. C)
backend/app/api/v1/endpoints/notifications_facebook_demo.py defined 8
endpoints (CreateNotification, BatchNotification, PreferenceUpdate, etc.)
that have NEVER been router-mounted in app/api/v1/api.py. The module
was unreachable in production; OpenAPI never knew about it; the frontend
never called it. Issue #665 categorised it as "dead module" (category C).

The only live consumer was the schema-contract test
backend/app/tests/test_facebook_notification_request_schemas.py
(Wave 6a112), which pins the 3 inline Pydantic request models for
PRIVACY/rate-limit defaults.

This PR:

1. Extracts the 3 Pydantic schemas (CreateNotificationRequest,
   BatchNotificationRequest, PreferenceUpdateRequest) into a clean
   schemas module: backend/app/schemas/notification_facebook.py.
   The schemas keep their existing defaults verbatim — the contract
   tests still pass without modification.
2. Repoints the test import to the new schemas module.
3. Updates the test module docstring to reflect the new home.
4. Deletes backend/app/api/v1/endpoints/notifications_facebook_demo.py
   entirely — including 8 endpoint handlers that were unreachable code.

Behaviour change for callers: none (no caller existed).

Code-quality wins:
- Removes 8 endpoint handlers that would never have been hit but were
  showing up in every observability audit / orphan scan.
- Schemas live where the rest of the project's schemas live
  (`app/schemas/`), not buried in a never-mounted endpoint file.

If the Facebook-style notification flow is ever resurrected, it should
land as a new router file importing from `app.schemas.notification_facebook`
and be deliberately mounted in `app/api/v1/api.py`.
